+++ back-end/tons-of-tacos/src/main/java/com/adamstraub/tonsoftacos/dto/customerDto/ordersDto/OrderReturnedToCustomer.java
@@ -3,6 +3,7 @@ package com.adamstraub.tonsoftacos.dto.customerDto.ordersDto;
 import com.adamstraub.tonsoftacos.dto.customerDto.orderItemsDto.OrderItemReturnedToCustomer;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -10,6 +11,7 @@ public class OrderReturnedToCustomer {
     private String customerName;
     private String orderUid;
     private List<OrderItemReturnedToCustomer> orderItems;
-    private Double orderTotal;
+//    private Double orderTotal;
+private BigDecimal orderTotal;
 
 }
